+++ tests/conftest.py
@@ -1,3 +1,4 @@
+"""Pytest fixtures and plugins."""
 from typing import Dict, Optional
 
 import logging
@@ -5,9 +6,6 @@ import os
 
 import pytest
 import yaml
-
-logger = logging.getLogger(__name__)
-
 
 LOG = logging.getLogger(__name__)
 
@@ -75,51 +73,6 @@ def _override_env_vars(overrides):
     # type: (Dict[str, Optional[str]]) -> Dict[str, str]
     """Use a dict to override os.environ values.
 
-    Args:
-        overrides: A dict of values to use for overrides.
-
-    Returns:
-        Copy of original os.environ values to be reinstated.
-
-    """
-    saved_env = {}
-    for key, value in overrides.items():
-        logger.info('Overriding env var: {}={}'.format(key, value))
-        saved_env[key] = os.environ.get(key, None)
-        os.environ[key] = value
-    return saved_env
-
-
-def _restore_env_vars(saved_env):
-    # type: (Dict[str, Optional[str]]) -> None
-    """Use a dict to revert os.environ values.
-
-    Args:
-        saved_env: A dict of values to restore.
-
-    """
-    for key, value in saved_env.items():
-        logger.info('Restoring saved env var: {}={}'.format(key, value))
-        if value is None:
-            del os.environ[key]
-        else:
-            os.environ[key] = value
-
-
-def _env_override(overrides):
-    # type: (Dict[str, str]) -> None
-    """Override environment variables for tests."""
-    saved_env = _override_env_vars(overrides)
-    yield
-    _restore_env_vars(saved_env)
-    saved_env.clear()
-
-
-@pytest.fixture(scope='session', autouse=True)
-def aws_credentials():
-    # type: () -> Dict[str, str]
-    """Handle change in https://github.com/spulec/moto/issues/1924.
-
     Ensure AWS SDK finds some (bogus) credentials in the environment and
     doesn't try to use other providers.
 
@@ -129,7 +82,22 @@ def aws_credentials():
         'AWS_SECRET_ACCESS_KEY': 'testing',
         'AWS_DEFAULT_REGION': 'us-east-1'
     }
-    return _env_override(overrides)
+    saved_env = {}
+    for key, value in overrides.items():
+        LOG.info('Overriding env var: {}={}'.format(key, value))
+        saved_env[key] = os.environ.get(key, None)
+        os.environ[key] = value
+
+    yield
+
+    for key, value in saved_env.items():
+        LOG.info('Restoring saved env var: {}={}'.format(key, value))
+        if value is None:
+            del os.environ[key]
+        else:
+            os.environ[key] = value
+
+    saved_env.clear()
 
 
 @pytest.fixture(scope='package')